Refactor fields into folder, add shared utils file
