+++ frontend/src/components/sampleTypes/fields/RemoveFieldBtn.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import IconButton from '@material-ui/core/IconButton';
+import CloseIcon from '@material-ui/icons/Close';
+import Grid from '@material-ui/core/Grid';
+
+const RemoveFieldBtn = ({ handleRemoveFieldClick, index }) => (
+  <Grid item lg={1} data-cy="remove-field-btn-div">
+    <IconButton
+      onClick={handleRemoveFieldClick(index)}
+      data-cy="remove-field-btn"
+    >
+      <CloseIcon />
+    </IconButton>
+  </Grid>
+);
+RemoveFieldBtn.propTypes = {
+  handleRemoveFieldClick: PropTypes.func.isRequired,
+  index: PropTypes.number.isRequired,
+};
+
+export default RemoveFieldBtn;
