+++ frontend/src/helpers/utils.js
@@ -1,0 +1,6 @@
+const randString = () => Math.random().toString(36).substr(7);
+const utils = {
+  randString,
+};
+
+export default utils;
